requirements.txt
farm-haystack[all]